Moving back to a single UMD bundle
