+++ package.json
@@ -10,8 +10,7 @@
   ],
   "version": "0.4.28",
   "files": [
-    "dist/teddy.js",
-    "dist/web.js"
+    "dist/teddy.js"
   ],
   "homepage": "https://github.com/rooseveltframework/teddy",
   "license": "CC-BY-4.0",
