+++ playground.html
@@ -39,7 +39,7 @@
     </main>
 
     <!-- include teddy library -->
-    <script src='dist/web.js'></script>
+    <script src='dist/teddy.js'></script>
 
     <script>
     (function() {
